Interimm commit - Minor progress towards Reuslt Pattern implementation for use cases

+++ ShiftLogger.Application/Shifts/Commands/DeleteShift/DeleteShiftHandler.cs
@@ -1,4 +1,6 @@
 ﻿using ShiftLogger.Domain.Models;
+using ShiftLogger.Domain.Validation;
+using ShiftLogger.Domain.Validation.Errors;
 using ShiftLogger.Infrastructure.Repositories;
 
 namespace ShiftLogger.Application.Shifts.Commands.DeleteShift;
@@ -12,8 +14,16 @@ public class DeleteShiftHandler
         _shiftsRepository = shiftsRepository;
     }
 
-    public async Task HandleAsync(DeleteShiftCommand command)
+    public async Task<Result> HandleAsync(DeleteShiftCommand command)
     {
+        var result = await _shiftsRepository.ShiftExistsById(command.Id);
+
+        if (result is null || result.Value == false)
+            return Result.Failure(Errors.ShiftNotFound);
+
+        if (result.IsFailiure)
+            return result;
+
         await _shiftsRepository.DeleteShiftAsync(new Shift
         {
             Id = command.Id,
@@ -22,6 +32,6 @@ public class DeleteShiftHandler
             ClockOutTime = command.ClockOutTime
         });
 
-        await _shiftsRepository.SaveChangesAsync();
+        return Result.Success();
     }
 }
